+++ Nest/Model/Domain/VrstaPostupka.cs
@@ -17,7 +17,7 @@ namespace Nest.Model.Domain {
         public virtual string Naziv { get; set; }
         public virtual string Opis { get; set; }
         [Required]
-        public virtual decimal Cijena { get; set; }
+        public virtual double Cijena { get; set; }
         public virtual bool Aktivno { get; set; }
         public virtual IEnumerable<Postupak> Postupaks { get; set; }
